+++ IcVibracoes.Core/Operations/RigidBody/CalculateVibration/TwoDegreesFreedom/ICalculateVibrationToTwoDegreesFreedom.cs
@@ -5,6 +5,6 @@ namespace IcVibracoes.Core.Operations.RigidBody.CalculateVibration.TwoDegreesFre
     /// <summary>
     /// It is responsible to calculate the vibration for a rigid body with two degrees freedom case.
     /// </summary>
-    public interface ICalculateVibrationToTwoDegreesFreedom : IOperationBase<TwoDegreesFreedomRequest, TwoDegreesFreedomResponse, TwoDegreesFreedomResponseData>
+    public interface ICalculateVibrationToTwoDegreesFreedom : ICalculateVibration_RigidBody<TwoDegreesFreedomRequest, TwoDegreesFreedomRequestData, TwoDegreesFreedomResponse, TwoDegreesFreedomResponseData>
     { }
 }